requirements.txt
bittensor==9.10.1
bt-decode==v0.6.0
bittensor-cli==9.11.0
async-substrate-interface==1.5.2
starlette>=0.30.0
pydantic>=2
rich>=13
pytest>=8
torch>=2
numpy>=1
setuptools>=68
httpx>=0.23.0
python-dotenv>=1.0.0
boto3>=1.26.0
wandb
sentence-transformers==3.4.1
SciPy==1.15.2
pytesseract